Added a `destroy()` method
It would be helpful to have a method that does the opposite of `setup()` for removing the sync hooks. The use case being I want local dev instances of sites to use the live Elasticsearch index but I don't want post updates/deletions on my local set-up to propagate to the production index.

Adding a method to this class would simplify things since when new hooks are added to ElasticPress, I wouldn't need to keep up with the changes. I could just call one method and be done with it.