return the right value for found_posts
When LIMITS are set in the query, we use the filter found_posts_query to return an empty string. $wpdb->get_var('') returns then the last query, which can be anything and put it in found_posts. Later in the set_found_posts function, when calculating the pagination, the function fails if found_posts as a rsult of the $wpdb->get_var('') was not numeric.

Using the filter found_posts, if the query was using limits, we can replace the found_posts value with the right one, as the number of posts in the cached query of Elasticpress