error_fix(SearchOrdering.php) - Invalid Argument
{"message":"Invalid argument supplied for foreach()","context":{"exception":{"class":"ErrorException","message":"Invalid argument supplied for foreach()","code":0,"file":"/var/www/html/pdinfo/releases/20190829162419/web/app/plugins/elasticpress/includes/classes/Feature/SearchOrdering/SearchOrdering.php:743"}

It is possible for the pointers variable in function handle_post_trash to be empty. This is generally not an issue unless running a very strict policy of not allowing any Errors. Our use case requires that we verify that variables being manipulated are verified as not empty.

Please consider accepting this change.