Ensure array query parameters do not contain empty items
In some cases query args like `post__in` that are expected to be array might contain invalid data such as an empty string. An array like `[ '' ]` will not appear as empty when checked with `empty( [ '' ] )` so this update ensures the arrays only contain real values. Database IDs all start from 1 so removing `0` is ok here.

This issue was observed with Yoast SEO adding subqueries for indexables to populate `post__in` parameters, however the index was incomplete. Can also happens in non-production environments where indexables aren't saved to the database.