Multisite context switch support to SyncManager
This commit adds support for Multisite context switching to the SyncManager class.
This allows it to index the content to the correct site when indexing in the shutdown hook.

The commit introduces two methods in the SyncManager class: `get_sync_queue` and `reset_sync_queue`.
The `get_sync_queue` method returns the sync queue for the current site or the `$blog_id`.
The `reset_sync_queue` method resets the sync queue for the current site or the `$blog_id`.

This commit also updates all the dependent tests to use the new methods.

Fixes #3471 issue.